Split binary negotiation tests into focused modules
