+++ crates/transport/src/binary/tests/mod.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+mod handshake;
+mod helpers;
+mod mapping;
+mod negotiation;
